Add files via upload
A university registrar’s office maintains data about the following entities:
1. courses, including number, title, credits, syllabus, and prerequisites;
2. course offerings, including course number, year, semester, section number,
instructor(s), timings, and classroom;
3. students, including student-id, name, and program;
4. instructors, including identi-cation number, name, department, and title.
Further, the enrollment of students in courses and grades awarded to students in each course they
are enrolled for must be appropriately modeled. Construct an E-R diagram for the registrar’s
of-ce.Document all assumptions that you make about the mapping constraints.